// apps/web/src/prismicio-types.d.ts
export type CalloutBoxSliceDefault = prismic.SharedSliceVariation<
  'default',
  Simplify<{
    type: prismic.SelectField<'tip' | 'warning' | 'info'>
    heading: prismic.KeyTextField
    content: prismic.RichTextField
  }>,
  never
>
export type CalloutBoxSliceVariation = CalloutBoxSliceDefault
export type CalloutBoxSlice = prismic.SharedSlice<'callout_box', CalloutBoxSliceVariation>